defer: Move more inlineCallbacks tests to correct module
